requirements.txt
ollama>=0.1.0
python-dotenv>=0.19.0
gradio>=4.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
pypdf>=3.0.0
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.2
langchain>=0.1.0
tiktoken>=0.5.0 